enhance prepare.sh ( stop on errors  )

+++ prepare.sh
@@ -18,11 +18,21 @@ cp $GEOBIN/config.xml .
 
 echo "get coastline data, to be able to draw seas."
 wget osmdata.openstreetmap.de/download/water-polygons-split-4326.zip
+if [ $? -ne 0]
+  then
+  echo " failed to download water polygons"
+  exit 1
+fi  
 unzip water-polygons-split-4326.zip -d .
 rm water-polygons-split-4326.zip
 
 echo "get countries boundaries from  natural earth"
 wget https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_boundary_lines_land.zip
+if [ $? -ne 0]
+  then
+  echo " failed to download boundaries from natural earth"
+  exit 1
+fi  
 #wget www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_boundary_lines_land.zip
 mkdir ne_10m_admin_0_boundary_lines_land
 unzip ne_10m_admin_0_boundary_lines_land.zip -d ne_10m_admin_0_boundary_lines_land
@@ -30,6 +40,11 @@ rm ne_10m_admin_0_boundary_lines_land.zip
 
 echo "get lakes from natural earth"
 wget https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/physical/ne_10m_lakes.zip
+if [ $? -ne 0]
+  then
+  echo " failed to download lakes from natural earth"
+  exit 1
+fi  
 #wget www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/physical/ne_10m_lakes.zip
 mkdir ne_10m_lakes
 unzip ne_10m_lakes.zip -d ne_10m_lakes
@@ -37,9 +52,19 @@ rm  ne_10m_lakes.zip
 
 echo "extract data from openstreetmap file"
 wget -O - $OSMFILE | tee osm.bz2 | lbzcat | $GEOBIN/renumber .
+if [ $? -ne 0]
+  then
+  echo " first data extraction step failed"
+  exit 1
+fi  
 
 echo "compile data from openstreetmap file and other shp files from natural earth"
 lbzcat osm.bz2 | $GEOBIN/compile .
+if [ $? -ne 0]
+  then
+  echo " failed to compile data"
+  exit 1
+fi  
 rm osm.bz2
 
 echo "remove indexes that are no longer useful"
@@ -49,6 +74,11 @@ rm relationIdIndex
 
 echo "create indexes"
 $GEOBIN/index .
+if [ $? -ne 0]
+  then
+  echo " index creation failed"
+  exit 1
+fi  
 
 echo "make cache directory"
 mkdir cache
